requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --annotation-style=line --output-file=requirements.txt requirements.in
#

anyio==3.6.1              # via starlette
certifi==2022.6.15        # via requests
charset-normalizer==2.1.0  # via requests
click==8.1.3              # via uvicorn
colorama==0.4.5           # via click
commonmark==0.9.1         # via rich
fastapi==0.79.0           # via -r requirements.in
h11==0.13.0               # via uvicorn
idna==3.3                 # via anyio, requests
pydantic==1.9.1           # via fastapi
pygments==2.12.0          # via rich
requests==2.28.1          # via -r requirements.in
rich==12.5.1              # via -r requirements.in
ruamel-yaml==0.17.21      # via -r requirements.in
ruamel-yaml-clib==0.2.6   # via ruamel-yaml
sniffio==1.2.0            # via anyio
starlette==0.19.1         # via -r requirements.in, fastapi
typing-extensions==4.3.0  # via pydantic, starlette
urllib3==1.26.10          # via requests
uvicorn==0.18.2           # via -r requirements.in
websockets==10.3          # via -r requirements.in
